Extract ConfigDatabaseSetDefaultCommand
